// src/interfaces/interfaces.ts
export interface Product {
  id: number
  name: string
  price: number
  image: string
  color: string
  quantity: number
}

export interface DeliveryInfo {
  fullName: string
  address: string
  city: string
  postalCode: string
  phone: string
  email: string
}

export interface Columns{
  header:string,
  accessorKey:string
}